Don't require the GUI thread's assistance to render video
The scene graph is able to render on its own thread and mpv is
notifying us of a new frame from another thread. This gives us
the opportunity to render video without asking if the GUI thread
has any QQuickItem to update, and avoid stuttering the video if
anything unexpected is blocking the GUI thread's event loop.

Possibly helps for #257