Check for local web client before downloading
When packaging for Ubuntu on Launchpad it's not allowed to download external resources as part of the build process. I believe this is true for other build systems too, but not 100% sure. This change would check the src folder for the appropriate version of the web client and if it's not there it would be downloaded just as before.

Plex-CLA-1.0-signed-off-by: Niclas Berglind <nb@kjam.se>