Probe ANGLE GLES version
The tricky part here is that ANGLE does not provide a way to do this
automatically. If you request 2, you get 2. If you request 3, and 3
can't be provided, it will fail. So we have to probe this manually.

This implies Qt has been patched to pass --disable-gpu to chromium.
(Because on Windows, chromium/QtWebEngine ignores the command line it is
given.)

Using GLES 3.x allows us to do higher quality video rendering. (Though
this also has to be manually enabled, and we don't provide an automatic
setting yet.)