Remove some more unnecessary scale things and fix default scale
It apears the web-client was still using the "system.scale" property,
which was hardcoded to return 0 with the recent scale code removal.
Change it to 1, which fixes HIDPI behavior here.

Passing the "initialScale" parameter to the web-client also seems to be
unnecessary now. web-client should stop parsing it, and should not read
"system.scale" anymore. For now, a stub is kept for "system.scale",
which always returns 1.