LinuxConfiguration: Install desktop files&icons
The desktop files (entries) are required to launch plexmediaplayer
via a GUI (instead of launching it via a terminal), the accompanying
icon is required so that the desktop entry has an icon.
Until now distros had to carry downstream patches to install
desktop files and plex's icons, with this commit we install
our own files.

Plex-CLA-1.0-signed-off-by: Rasmus Thomsen <cogitri@exherbo.org>