powercomponent: only report relaunch capability on embedded, fixes #354
Reporting the relaunch capability gets web to display a relaunch menu entry.

That menu entry is only valid on Embedded where systemd will keep on relaunching upon PMP exit.
On linux builds, this will get reported as supported, and will therefore show the relaunch menu item.
When users click on it, it will simply exit, without relaunching.

We just make it valid for embedded with that patch.