icon: Do not enforce icon on startup, fixes #622
We enforce the icon onstartup and this has been reported as an issue for people willing to change the application Icon on OSX.

This is only required on Windows.
On OSX this would break ability of user to change the App icon which is made from the SVG version.
We limit setting this window icon to Windows then.